feat: show proper images on each onboarding screen
